ai_screenwriter_tool/tests/test_project.py
```python
import unittest
from ai_screenwriter_tool.project import Project

class TestProject(unittest.TestCase):

    def setUp(self):
        self.project = Project()

    def test_updateProject(self):
        self.project.updateProject('New Project', 'This is a new project.')
        self.assertEqual(self.project.name, 'New Project')
        self.assertEqual(self.project.description, 'This is a new project.')

    def test_generateAIInsights(self):
        insights = self.project.generateAIInsights()
        self.assertIsNotNone(insights)

    def test_provideBrainstormingAssistance(self):
        assistance = self.project.provideBrainstormingAssistance()
        self.assertIsNotNone(assistance)

    def test_overcomeWritersBlock(self):
        solution = self.project.overcomeWritersBlock()
        self.assertIsNotNone(solution)

    def test_applyEditingFeatures(self):
        self.project.applyEditingFeatures()
        self.assertTrue(self.project.edited)

    def test_seekExpertSupport(self):
        support = self.project.seekExpertSupport()
        self.assertIsNotNone(support)

    def test_enableCollaboration(self):
        self.project.enableCollaboration()
        self.assertTrue(self.project.collaboration_enabled)

if __name__ == '__main__':
    unittest.main()
```